update report page
