+++ Frontend/src/api/ai.jsx
@@ -64,7 +64,7 @@ export function FluxMessageWithHistoryAPI(params) {
 export function UploadFileAPI(formData) {
     console.log("Uploading file data:", formData)
     return request({
-        url: '/document/embedding',
+        url: '/etl/embedding/multipart',
         method: 'POST',
         headers: {
             'Content-Type': 'multipart/form-data'
