+++ Frontend/src/pages/report/index.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useRef, useEffect, useCallback } from 'react';
+import React, { useState, useRef, useCallback } from 'react';
 import {
   Box,
   Typography,
@@ -9,10 +9,6 @@ import {
   Grid,
   Switch,
   FormControlLabel,
-  List,
-  ListItem,
-  ListItemText,
-  ListItemIcon,
   IconButton,
   useTheme,
   alpha,
@@ -22,125 +18,49 @@ import {
   DialogContent,
   DialogActions,
   Tooltip,
-  Paper,
-  Chip,
-  Avatar,
 } from '@mui/material';
 import {
   Send as SendIcon,
-  Upload as UploadIcon,
-  InsertDriveFile as FileIcon,
-  Add as AddIcon,
-  Chat as ChatIcon,
-  Delete as DeleteIcon,
-  Edit as EditIcon,
-  Person as PersonIcon,
-  SmartToy as AIIcon,
   Menu as MenuIcon,
-  Refresh as RefreshIcon,
-  ContentCopy as CopyIcon,
-  Download as DownloadIcon,
 } from '@mui/icons-material';
-import { v4 as uuidv4 } from 'uuid';
-import { motion, AnimatePresence } from 'framer-motion';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import rehypeRaw from 'rehype-raw';
-import remarkMath from 'remark-math';
-import rehypeKatex from 'rehype-katex';
-import 'katex/dist/katex.min.css';
-import {
-  MessageAPI,
-  FluxMessageAPI,
-  FluxMessageWithHistoryAPI,
-  ChatWithFileAPI,
-  UploadFileAPI
-} from '@/api/ai';
+import { useChatSessions } from '@/hooks/useChatSessions';
+import { useFileUpload } from '@/hooks/useFileUpload';
+import { Sidebar } from '@/components/Sidebar';
+import { ChatMessages } from '@/components/ChatMessages';
+import { FluxMessageWithHistoryAPI, ChatWithFileAPI } from '@/api/ai';
+
+const drawerWidth = 240;
 
 export default function ChatInterface() {
   const theme = useTheme();
-  const [sessions, setSessions] = useState(() => {
-    try {
-      const savedSessions = localStorage.getItem('chatSessions');
-      return savedSessions ? JSON.parse(savedSessions) : [{ id: uuidv4(), name: 'New Chat', messages: [] }];
-    } catch (error) {
-      console.error('Error loading sessions from localStorage:', error);
-      return [{ id: uuidv4(), name: 'New Chat', messages: [] }];
-    }
-  });
-  const [activeSession, setActiveSession] = useState(() => sessions[0]?.id || '');
+  const {
+    sessions,
+    activeSession,
+    setActiveSession,
+    addNewSession,
+    deleteSession,
+    updateSessionName,
+    addMessageToActiveSession,
+  } = useChatSessions();
+  const { files, uploadFile, clearFiles } = useFileUpload();
+
   const [message, setMessage] = useState('');
   const [isRetrievalMode, setIsRetrievalMode] = useState(false);
-  const [files, setFiles] = useState(() => {
-    try {
-      const savedFiles = localStorage.getItem('uploadedFiles');
-      return savedFiles ? JSON.parse(savedFiles) : [];
-    } catch (error) {
-      console.error('Error loading files from localStorage:', error);
-      return [];
-    }
-  });
-  const [username, setUsername] = useState(() => localStorage.getItem('username') || 'User');
   const [isLoading, setIsLoading] = useState(false);
   const [newSessionDialogOpen, setNewSessionDialogOpen] = useState(false);
   const [newSessionName, setNewSessionName] = useState('');
   const [editSessionId, setEditSessionId] = useState(null);
   const fileInputRef = useRef(null);
-  const chatContainerRef = useRef(null);
   const [isTyping, setIsTyping] = useState(false);
   const [sidebarOpen, setSidebarOpen] = useState(true);
-
-  useEffect(() => {
-    try {
-      localStorage.setItem('chatSessions', JSON.stringify(sessions));
-    } catch (error) {
-      console.error('Error saving sessions to localStorage:', error);
-    }
-  }, [sessions]);
-
-  useEffect(() => {
-    try {
-      localStorage.setItem('uploadedFiles', JSON.stringify(files));
-    } catch (error) {
-      console.error('Error saving files to localStorage:', error);
-    }
-  }, [files]);
-
-  useEffect(() => {
-    if (chatContainerRef.current) {
-      chatContainerRef.current.scrollTop = chatContainerRef.current.scrollHeight;
-    }
-  }, [sessions]);
-
-  const isChineseChar = (char) => {
-    const charCode = char.charCodeAt(0);
-    return (charCode >= 0x4e00 && charCode <= 0x9fff) ||
-        (charCode >= 0x3400 && charCode <= 0x4dbf) ||
-        (charCode >= 0x20000 && charCode <= 0x2a6df) ||
-        (charCode >= 0x2a700 && charCode <= 0x2b73f) ||
-        (charCode >= 0x2b740 && charCode <= 0x2b81f) ||
-        (charCode >= 0x2b820 && charCode <= 0x2ceaf);
-  };
-
-  const isChinesePunctuation = (char) => {
-    const chinesePunctuationRegex = /[\u3000-\u303f\uff00-\uff0f\uff1a-\uff20\uff3b-\uff40\uff5b-\uff65]/;
-    return chinesePunctuationRegex.test(char);
-  };
+  const [username, setUsername] = useState(() => localStorage.getItem('username') || 'User');
 
   const handleSendMessage = useCallback(async () => {
     if (message.trim()) {
       const decodedMessage = decodeURIComponent(message.trim());
       console.log("User input:", decodedMessage);
-      const updatedSessions = [...sessions];
-      const sessionIndex = updatedSessions.findIndex(s => s.id === activeSession);
 
-      if (sessionIndex === -1) {
-        console.error('Active session not found');
-        return;
-      }
-
-      updatedSessions[sessionIndex].messages.push({ sender: username, content: decodedMessage });
-      setSessions(updatedSessions);
+      addMessageToActiveSession({ sender: username, content: decodedMessage });
       setMessage('');
       setIsLoading(true);
       setIsTyping(true);
@@ -165,113 +85,41 @@ export default function ChatInterface() {
         const sseData = response.data;
         const lines = sseData.split('\n');
         let aiResponse = '';
-        let currentLine = '';
-        let isChineseResponse = false;
-
-        updatedSessions[sessionIndex].messages.push({ sender: 'AI', content: '' });
-        setSessions([...updatedSessions]);
 
         for (const line of lines) {
           if (line.startsWith('data:')) {
             const messagePart = line.replace('data:', '').trim();
-            if (messagePart === '') {
-              // Empty data, treat as a new line
-              aiResponse += currentLine + '\n';
-              currentLine = '';
-            } else {
-              if (currentLine === '' && isChineseChar(messagePart[0])) {
-                isChineseResponse = true;
-              }
-
-              if (isChineseResponse) {
-                // For Chinese, add characters directly and start a new line after punctuation
-                for (const char of messagePart) {
-                  currentLine += char;
-                  if (isChinesePunctuation(char)) {
-                    aiResponse += currentLine + '\n';
-                    currentLine = '';
-                  }
-                }
-              } else {
-                // For English, add space before the new word, unless it's the first word in the line
-                currentLine += (currentLine ? ' ' : '') + messagePart;
-              }
-            }
-            updatedSessions[sessionIndex].messages[updatedSessions[sessionIndex].messages.length - 1] = {
-              sender: 'AI',
-              content: aiResponse + currentLine
-            };
-            setSessions([...updatedSessions]);
+            aiResponse += messagePart;
+            addMessageToActiveSession({ sender: 'AI', content: aiResponse });
             await new Promise(resolve => setTimeout(resolve, 50));
           }
         }
-        // Add the last line if it's not empty
-        if (currentLine) {
-          aiResponse += currentLine;
-          updatedSessions[sessionIndex].messages[updatedSessions[sessionIndex].messages.length - 1] = {
-            sender: 'AI',
-            content: aiResponse
-          };
-          setSessions([...updatedSessions]);
-        }
       } catch (error) {
         console.error('Error sending message:', error);
-        updatedSessions[sessionIndex].messages.push({ sender: 'AI', content: 'Sorry, there was an error processing your request.' });
-        setSessions([...updatedSessions]);
+        addMessageToActiveSession({ sender: 'AI', content: 'Sorry, there was an error processing your request.' });
       } finally {
         setIsLoading(false);
         setIsTyping(false);
       }
     }
-  }, [message, sessions, activeSession, isRetrievalMode, files, username]);
+  }, [message, activeSession, isRetrievalMode, files, username, addMessageToActiveSession]);
 
   const handleFileUpload = async (event) => {
-    const selectedFile = event.target.files[0];
+    const selectedFile = event.target.files?.[0];
     if (selectedFile) {
       setIsLoading(true);
       try {
-        const formData = new FormData();
-        formData.append('file', selectedFile);
-        const response = await UploadFileAPI(formData);
-        if (response.status === 200) {
-          const fileMetadata = {
-            name: selectedFile.name,
-            lastModified: selectedFile.lastModified,
-            size: selectedFile.size,
-            type: selectedFile.type
-          };
-          setFiles(prev => [...prev, fileMetadata]);
-        } else {
-          throw new Error('File upload failed');
-        }
-      } catch (error) {
-        console.error('Error uploading file:', error);
-        // You might want to show an error message to the user here
+        await uploadFile(selectedFile);
       } finally {
         setIsLoading(false);
       }
     }
   };
 
-  const addNewSession = () => {
-    setNewSessionDialogOpen(true);
-  };
-
   const handleNewSessionConfirm = () => {
-    const newSession = { id: uuidv4(), name: newSessionName || `New Chat ${sessions.length + 1}`, messages: [] };
-    setSessions(prev => [...prev, newSession]);
-    setActiveSession(newSession.id);
+    addNewSession(newSessionName);
     setNewSessionDialogOpen(false);
     setNewSessionName('');
-  };
-
-  const deleteSession = (id) => {
-    if (sessions.length > 1) {
-      setSessions(prev => prev.filter(s => s.id !== id));
-      if (activeSession === id) {
-        setActiveSession(sessions.find(s => s.id !== id)?.id);
-      }
-    }
   };
 
   const startEditSession = (id) => {
@@ -280,11 +128,11 @@ export default function ChatInterface() {
   };
 
   const handleEditSessionConfirm = () => {
-    setSessions(prev => prev.map(s =>
-        s.id === editSessionId ? { ...s, name: newSessionName } : s
-    ));
-    setEditSessionId(null);
-    setNewSessionName('');
+    if (editSessionId) {
+      updateSessionName(editSessionId, newSessionName);
+      setEditSessionId(null);
+      setNewSessionName('');
+    }
   };
 
   const toggleSidebar = () => {
@@ -292,16 +140,15 @@ export default function ChatInterface() {
   };
 
   const handleRetry = async () => {
-    const sessionIndex = sessions.findIndex(s => s.id === activeSession);
-    if (sessionIndex === -1) return;
+    const currentSession = sessions.find(s => s.id === activeSession);
+    if (!currentSession) return;
 
-    const lastUserMessage = [...sessions[sessionIndex].messages].reverse().find(m => m.sender === username);
+    const lastUserMessage = [...currentSession.messages].reverse().find(m => m.sender === username);
     if (!lastUserMessage) return;
 
     // Remove the last AI message
-    const updatedSessions = [...sessions];
-    updatedSessions[sessionIndex].messages.pop();
-    setSessions(updatedSessions);
+    const updatedMessages = currentSession.messages.slice(0, -1);
+    updateSessionName(activeSession, currentSession.name);
 
     // Resend the last user message
     setMessage(lastUserMessage.content);
@@ -310,7 +157,6 @@ export default function ChatInterface() {
 
   const handleCopy = (content) => {
     navigator.clipboard.writeText(content).then(() => {
-      // You might want to show a success message here
       console.log('Content copied to clipboard');
     }, (err) => {
       console.error('Could not copy text: ', err);
@@ -327,210 +173,33 @@ export default function ChatInterface() {
     document.body.removeChild(element);
   };
 
-  const renderMessage = (content, sender) => {
-    const MarkdownComponents = {
-      p: ({ node, ...props }) => <Typography variant="body1" paragraph sx={{ wordBreak: 'break-word', whiteSpace: 'pre-wrap' }} {...props} />,
-      h1: ({ node, ...props }) => <Typography variant="h4" gutterBottom sx={{ wordBreak: 'break-word', mt: 4, mb: 2 }} {...props} />,
-      h2: ({ node, ...props }) => <Typography variant="h5" gutterBottom sx={{ wordBreak: 'break-word', mt: 3, mb: 2 }} {...props} />,
-      h3: ({ node, ...props }) => <Typography variant="h6" gutterBottom sx={{ wordBreak: 'break-word', mt: 2, mb: 1 }} {...props} />,
-      h4: ({ node, ...props }) => <Typography variant="subtitle1" gutterBottom sx={{ wordBreak: 'break-word', fontWeight: 'bold' }} {...props} />,
-      h5: ({ node, ...props }) => <Typography variant="subtitle2" gutterBottom sx={{ wordBreak: 'break-word', fontWeight: 'bold' }} {...props} />,
-      h6: ({ node, ...props }) => <Typography variant="subtitle2" gutterBottom sx={{ wordBreak: 'break-word', fontStyle: 'italic' }} {...props} />,
-      ul: ({ node, ...props }) => <ul style={{ paddingLeft: '20px', marginBottom: '16px' }} {...props} />,
-      ol: ({ node, ...props }) => <ol style={{ paddingLeft: '20px', marginBottom: '16px' }} {...props} />,
-      li: ({ node, ...props }) => <li style={{ marginBottom: '8px' }} {...props} />,
-      code: ({ node, inline, ...props }) =>
-          inline ? (
-              <code style={{ backgroundColor: alpha(theme.palette.primary.main, 0.1), padding: '2px 4px', borderRadius: '4px', wordBreak: 'break-word' }} {...props} />
-          ) : (
-              <pre style={{ backgroundColor: alpha(theme.palette.primary.main, 0.1), padding: '16px', borderRadius: '4px', overflowX: 'auto', whiteSpace: 'pre-wrap', wordBreak: 'break-word' }}>
-            <code {...props} />
-          </pre>
-          ),
-      strong: ({ node, ...props }) => <strong style={{ fontWeight: 'bold' }} {...props} />,
-    };
-
-    if (sender === 'AI') {
-      // Process AI response
-      const processedContent = content
-          .split('\n')
-          .map(line => line.trim())
-          .join('\n')
-          .replace(/\n{3,}/g, '\n\n'); // Replace 3 or more consecutive newlines with 2
-
-      return (
-          <>
-            <ReactMarkdown
-                remarkPlugins={[remarkGfm, remarkMath]}
-                rehypePlugins={[rehypeRaw, rehypeKatex]}
-                components={MarkdownComponents}
-            >
-              {processedContent}
-            </ReactMarkdown>
-            <Box sx={{ display: 'flex', justifyContent: 'flex-end', mt: 1 }}>
-              <IconButton onClick={() => handleRetry()} size="small">
-                <RefreshIcon fontSize="small" />
-              </IconButton>
-              <IconButton onClick={() => handleCopy(processedContent)} size="small">
-                <CopyIcon fontSize="small" />
-              </IconButton>
-              <IconButton onClick={() => handleDownload(processedContent)} size="small">
-                <DownloadIcon fontSize="small" />
-              </IconButton>
-            </Box>
-          </>
-      );
-    } else {
-      // For user messages, maintain existing behavior
-      return (
-          <Typography
-              variant="body1"
-              sx={{
-                color: theme.palette.text.primary,
-                whiteSpace: 'pre-wrap',
-                wordBreak: 'break-word',
-              }}
-          >
-            {content}
-          </Typography>
-      );
-    }
-  };
-
   return (
-      <Box sx={{ display: 'flex', height: '95vh', overflow: 'hidden' }}>
+      <Box sx={{ display: 'flex', height: '100vh', overflow: 'hidden' }}>
         <Box sx={{
-          width: sidebarOpen ? 240 : 0,
+          width: sidebarOpen ? drawerWidth : 0,
           flexShrink: 0,
           transition: theme.transitions.create('width', {
             easing: theme.transitions.easing.sharp,
             duration: theme.transitions.duration.enteringScreen,
           }),
         }}>
-          <Paper
-              elevation={0}
-              sx={{
-                height: '100%',
-                overflow: 'auto',
-                borderRadius: 0,
-                borderRight: 1,
-                borderColor: 'divider',
-              }}
-          >
-            <Box sx={{ p: 2, bgcolor: 'primary.main', color: 'primary.contrastText' }}>
-              <Typography variant="h6" sx={{ fontWeight: 'bold' }}>Chat Sessions</Typography>
-            </Box>
-            <List sx={{ flexGrow: 1 }}>
-              <AnimatePresence>
-                {sessions.map((session) => (
-                    <motion.div
-                        key={session.id}
-                        initial={{ opacity: 0, y: -20 }}
-                        animate={{ opacity: 1, y: 0 }}
-                        exit={{ opacity: 0, y: -20 }}
-                        transition={{ duration: 0.2 }}
-                    >
-                      <ListItem
-                          button
-                          selected={session.id === activeSession}
-                          onClick={() => setActiveSession(session.id)}
-                          sx={{
-                            borderRadius: 1,
-                            my: 0.5,
-                            mx: 1,
-                            '&.Mui-selected': {
-                              bgcolor: alpha(theme.palette.primary.main, 0.1),
-                              '&:hover': {
-                                bgcolor: alpha(theme.palette.primary.main, 0.2),
-                              }
-                            }
-                          }}
-                      >
-                        <ListItemIcon>
-                          <ChatIcon color={session.id === activeSession ? 'primary' : 'inherit'} />
-                        </ListItemIcon>
-                        {editSessionId === session.id ? (
-                            <TextField
-                                value={newSessionName}
-                                onChange={(e) => setNewSessionName(e.target.value)}
-                                onBlur={handleEditSessionConfirm}
-                                onKeyPress={(e) => {
-                                  if (e.key === 'Enter') {
-                                    handleEditSessionConfirm();
-                                  }
-                                }}
-                                autoFocus
-                                size="small"
-                                sx={{ flexGrow: 1 }}
-                            />
-                        ) : (
-                            <ListItemText primary={session.name} />
-                        )}
-                        <IconButton edge="end" onClick={() => startEditSession(session.id)} size="small">
-                          <EditIcon fontSize="small" />
-                        </IconButton>
-                        {sessions.length > 1 && (
-                            <IconButton edge="end" onClick={() => deleteSession(session.id)} size="small">
-                              <DeleteIcon fontSize="small" />
-                            </IconButton>
-                        )}
-                      </ListItem>
-                    </motion.div>
-                ))}
-              </AnimatePresence>
-            </List>
-            <Divider />
-            <Box sx={{ p: 2 }}>
-              <Button
-                  fullWidth
-                  variant="contained"
-                  startIcon={<AddIcon />}
-                  onClick={addNewSession}
-                  sx={{ borderRadius: 2 }}
-              >
-                New Chat
-              </Button>
-            </Box>
-            {isRetrievalMode && (
-                <>
-                  <Divider />
-                  <Box sx={{ p: 2 }}>
-                    <Typography variant="subtitle2" gutterBottom>Uploaded Files</Typography>
-                    <List dense>
-                      {files.map((file, index) => (
-                          <ListItem key={index}>
-                            <ListItemIcon>
-                              <FileIcon fontSize="small" />
-                            </ListItemIcon>
-                            <ListItemText
-                                primary={file.name}
-                                secondary={`${(file.size / 1024).toFixed(2)} KB`}
-                                primaryTypographyProps={{ variant: 'body2' }}
-                                secondaryTypographyProps={{ variant: 'caption' }}
-                            />
-                          </ListItem>
-                      ))}
-                    </List>
-                    <input
-                        type="file"
-                        onChange={handleFileUpload}
-                        style={{ display: 'none' }}
-                        ref={fileInputRef}
-                    />
-                    <Button
-                        fullWidth
-                        variant="outlined"
-                        startIcon={<UploadIcon />}
-                        onClick={() => fileInputRef.current?.click()}
-                        disabled={isLoading}
-                        sx={{ mt: 1, borderRadius: 2 }}
-                    >
-                      Upload File
-                    </Button>
-                  </Box>
-                </>
-            )}
-          </Paper>
+          <Sidebar
+              sessions={sessions}
+              activeSession={activeSession}
+              setActiveSession={setActiveSession}
+              addNewSession={() => setNewSessionDialogOpen(true)}
+              deleteSession={deleteSession}
+              startEditSession={startEditSession}
+              editSessionId={editSessionId}
+              newSessionName={newSessionName}
+              setNewSessionName={setNewSessionName}
+              handleEditSessionConfirm={handleEditSessionConfirm}
+              isRetrievalMode={isRetrievalMode}
+              files={files}
+              handleFileUpload={handleFileUpload}
+              fileInputRef={fileInputRef}
+              isLoading={isLoading}
+          />
         </Box>
         <Box sx={{
           flexGrow: 1,
@@ -539,9 +208,9 @@ export default function ChatInterface() {
           display: 'flex',
           flexDirection: 'column',
         }}>
-          <Box sx={{ p: 1.5, display: 'flex', justifyContent: 'space-between', alignItems: 'center', borderBottom: 1, borderColor: 'divider' }}>
+          <Box sx={{ p: 2, display: 'flex', justifyContent: 'space-between', alignItems: 'center', borderBottom: 1, borderColor: 'divider' }}>
             <Box sx={{ display: 'flex', alignItems: 'center' }}>
-              <IconButton onClick={toggleSidebar} sx={{ mr: 1 }}>
+              <IconButton onClick={toggleSidebar} sx={{ mr: 2 }}>
                 <MenuIcon />
               </IconButton>
               <Typography variant="h6" component="h1" sx={{ fontWeight: 'bold' }}>
@@ -555,7 +224,7 @@ export default function ChatInterface() {
                       onChange={(e) => {
                         setIsRetrievalMode(e.target.checked);
                         if (!e.target.checked) {
-                          setFiles([]);
+                          clearFiles();
                           if (fileInputRef.current) fileInputRef.current.value = '';
                         }
                       }}
@@ -564,66 +233,27 @@ export default function ChatInterface() {
                 label="Retrieval Mode"
             />
           </Box>
-          <Box sx={{ flexGrow: 1, overflow: 'hidden', display: 'flex', flexDirection: 'column', p: 1.5 }}>
-            <Card sx={{
-              flexGrow: 1,
-              display: 'flex',
-              flexDirection: 'column',
-              boxShadow: theme.shadows[4],
-              borderRadius: 2,
-              overflow: 'hidden',
-            }}>
-              <CardContent sx={{ flexGrow: 1, overflow: 'auto', p: 2 }} ref={chatContainerRef}>
-                <AnimatePresence>
-                  {sessions.find(s => s.id === activeSession)?.messages.map((msg, index) => (
-                      <motion.div
-                          key={index}
-                          initial={{ opacity: 0, y: 20 }}
-                          animate={{ opacity: 1, y: 0 }}
-                          exit={{ opacity: 0, y: -20 }}
-                          transition={{ duration: 0.2 }}
-                      >
-                        <Box sx={{ mb: 1.5, display: 'flex', justifyContent: msg.sender === username ? 'flex-end' : 'flex-start' }}>
-                          <Paper elevation={1} sx={{
-                            maxWidth: '70%',
-                            p: 1.5,
-                            borderRadius: 2,
-                            bgcolor: msg.sender === username
-                                ? alpha(theme.palette.primary.main, 0.1)
-                                : theme.palette.mode === 'dark'
-                                    ? alpha(theme.palette.background.paper, 0.2)
-                                    : '#f7f7f8',
-                          }}>
-                            <Box sx={{ display: 'flex', alignItems: 'center', mb: 0.5 }}>
-                              <Avatar sx={{ width: 24, height: 24, mr: 1, bgcolor: msg.sender === username ? 'primary.main' : 'secondary.main' }}>
-                                {msg.sender === username ? <PersonIcon fontSize="small" /> : <AIIcon fontSize="small" />}
-                              </Avatar>
-                              <Typography variant="subtitle2" sx={{ fontWeight: 'bold' }}>
-                                {msg.sender === username ? username : 'AI'}
-                              </Typography>
-                            </Box>
-                            {renderMessage(msg.content, msg.sender)}
-                          </Paper>
-                        </Box>
-                      </motion.div>
-                  ))}
-                </AnimatePresence>
-                {isTyping && (
-                    <Box sx={{ display: 'flex', justifyContent: 'flex-start', mb: 1.5 }}>
-                      <Chip
-                          icon={<AIIcon />}
-                          label="AI is typing..."
-                          variant="outlined"
-                          color="secondary"
-                          size="small"
-                          sx={{ animation: 'pulse 1.5s infinite' }}
-                      />
-                    </Box>
-                )}
-              </CardContent>
+          <Card sx={{
+            flexGrow: 1,
+            display: 'flex',
+            flexDirection: 'column',
+            boxShadow: theme.shadows[4],
+            borderRadius: 2,
+            overflow: 'hidden',
+            m: 2,
+          }}>
+            <CardContent sx={{ flexGrow: 1, overflow: 'hidden', display: 'flex', flexDirection: 'column', p: 0 }}>
+              <ChatMessages
+                  messages={sessions.find(s => s.id === activeSession)?.messages || []}
+                  username={username}
+                  isTyping={isTyping}
+                  handleRetry={handleRetry}
+                  handleCopy={handleCopy}
+                  handleDownload={handleDownload}
+              />
               <Divider />
-              <Box sx={{ p: 1.5 }}>
-                <Grid container spacing={1.5} alignItems="center">
+              <Box sx={{ p: 2 }}>
+                <Grid container spacing={2} alignItems="center">
                   <Grid item xs>
                     <TextField
                         fullWidth
@@ -638,7 +268,7 @@ export default function ChatInterface() {
                           }
                         }}
                         multiline
-                        maxRows={3}
+                        maxRows={4}
                         disabled={isLoading}
                         sx={{
                           '& .MuiOutlinedInput-root': {
@@ -664,8 +294,8 @@ export default function ChatInterface() {
                           disabled={isLoading || (isRetrievalMode && files.length === 0) || !message.trim()}
                           sx={{
                             borderRadius: 2,
-                            px: 2,
-                            py: 1,
+                            px: 3,
+                            py: 1.5,
                             transition: 'all 0.3s',
                             '&:hover': {
                               transform: 'translateY(-2px)',
@@ -680,8 +310,8 @@ export default function ChatInterface() {
                   </Grid>
                 </Grid>
               </Box>
-            </Card>
-          </Box>
+            </CardContent>
+          </Card>
         </Box>
 
         <Dialog
